feat: Script to monitor free space on a system
A bash script to monitor the amount of free space on a system.
The script sets default values for wait interval and threshold,
but also allows overide them using command line options.
The script checks the free space on the system using the `df` command
and compares it to the specified threshold value.
If the free space is less than the threshold, the script prints a warning message
and returns a false value.
The script can be run once or indefinitely based on the -i flag.